+++ app/SML/models/types/device/supportdevice.h
@@ -6,7 +6,12 @@
 class SupportDevice : public Device
 {
 public:
-    SupportDevice(QString name, QString label, QString index, bool activeState, int mask, QObject *parent = nullptr);
+    SupportDevice(QString name,
+                  QString label,
+                  QString index,
+                  bool activeState,
+                  byte mask,
+                  QObject *parent = nullptr);
     QStringList getParams() override;
 };
